fix all tests for windows & python 2.7/3.5

+++ tests/test_file_recognizer.py
@@ -8,9 +8,11 @@ import os
 import shutil
 import socket
 import sys
+import tempfile
 
 import nose
 import nose.tools as nt
+from unittest import skipIf
 
 from grin import FileRecognizer, GZIP_MAGIC, ALLBYTES
 
@@ -22,15 +24,25 @@ def empty_file(filename, open=open):
 
 def binary_file(filename, open=open):
     with open(filename, 'wb') as f:
-        f.write(bytes(range(256)))
+        f.write(bytes(bytearray(range(256))))
 
 
 def text_file(filename, open=open):
     lines = ['foo\n', 'bar\n'] * 100
     lines.append('baz\n')
     lines.extend(['foo\n', 'bar\n'] * 100)
+    text = "".join(lines)
     with open(filename, 'w') as f:
-        f.writelines(lines)
+        f.write(text)
+
+
+def text_file_gz(filename):
+    lines = ['foo\n', 'bar\n'] * 100
+    lines.append('baz\n')
+    lines.extend(['foo\n', 'bar\n'] * 100)
+    text = "".join(lines)
+    with gzip.open(filename, 'w') as f:
+        f.write(text.encode('utf8'))
 
 
 def fake_gzip_file(filename, open=open):
@@ -84,8 +96,17 @@ def totally_unusable_dir(filename):
     os.mkdir(filename)
     os.chmod(filename, 0o100)
 
+_temp_test_dir = None
+_saved_working_dir = None
+
 
 def setup():
+    global _temp_test_dir
+    global _saved_working_dir
+    _saved_working_dir = os.getcwd()
+    _temp_test_dir = tempfile.mkdtemp(prefix='grin-')
+    os.chdir(_temp_test_dir)
+
     # Make files to test individual recognizers.
     empty_file('empty')
     binary_file('binary')
@@ -99,31 +120,36 @@ def setup():
     text_file('.text')
     empty_file('empty.gz', open=gzip.open)
     binary_file('binary.gz', open=gzip.open)
-    text_file('text.gz', open=gzip.open)
+    text_file_gz('text.gz')
     binary_file('.binary.gz', open=gzip.open)
-    text_file('.text.gz', open=gzip.open)
+    text_file_gz('.text.gz')
     fake_gzip_file('fake.gz')
     os.mkdir('.dir')
-    os.symlink('binary', 'binary_link')
-    os.symlink('text', 'text_link')
-    os.symlink('dir', 'dir_link')
-    os.symlink('.binary', '.binary_link')
-    os.symlink('.text', '.text_link')
-    os.symlink('.dir', '.dir_link')
-    unreadable_file('unreadable_file')
-    unreadable_dir('unreadable_dir')
-    unexecutable_dir('unexecutable_dir')
-    totally_unusable_dir('totally_unusable_dir')
-    os.symlink('unreadable_file', 'unreadable_file_link')
-    os.symlink('unreadable_dir', 'unreadable_dir_link')
-    os.symlink('unexecutable_dir', 'unexecutable_dir_link')
-    os.symlink('totally_unusable_dir', 'totally_unusable_dir_link')
+    if not sys.platform == 'win32' or sys.version_info > (3, 2):
+        os.symlink('binary', 'binary_link')
+        os.symlink('text', 'text_link')
+        os.symlink('dir', 'dir_link')
+        os.symlink('.binary', '.binary_link')
+        os.symlink('.text', '.text_link')
+        os.symlink('.dir', '.dir_link')
+
+    if not sys.platform == 'win32':
+        unreadable_file('unreadable_file')
+        unreadable_dir('unreadable_dir')
+        unexecutable_dir('unexecutable_dir')
+        totally_unusable_dir('totally_unusable_dir')
+        os.symlink('unreadable_file', 'unreadable_file_link')
+        os.symlink('unreadable_dir', 'unreadable_dir_link')
+        os.symlink('unexecutable_dir', 'unexecutable_dir_link')
+        os.symlink('totally_unusable_dir', 'totally_unusable_dir_link')
     text_file('text.skip_ext')
     os.mkdir('dir.skip_ext')
     text_file('text.dont_skip_ext')
     os.mkdir('skip_dir')
     text_file('fake_skip_dir')
-    socket_file('socket_test')
+
+    if not sys.platform == 'win32':
+        socket_file('socket_test')
 
     # Make a directory tree to test tree-walking.
     os.mkdir('tree')
@@ -142,15 +168,16 @@ def setup():
     text_file('tree/skip_dir/text')
     os.mkdir('tree/.skip_hidden_dir')
     text_file('tree/.skip_hidden_file')
-    os.mkdir('tree/unreadable_dir')
-    text_file('tree/unreadable_dir/text')
-    os.chmod('tree/unreadable_dir', 0o300)
-    os.mkdir('tree/unexecutable_dir')
-    text_file('tree/unexecutable_dir/text')
-    os.chmod('tree/unexecutable_dir', 0o600)
-    os.mkdir('tree/totally_unusable_dir')
-    text_file('tree/totally_unusable_dir/text')
-    os.chmod('tree/totally_unusable_dir', 0o100)
+    if not sys.platform == 'win32':
+        os.mkdir('tree/unreadable_dir')
+        text_file('tree/unreadable_dir/text')
+        os.chmod('tree/unreadable_dir', 0o300)
+        os.mkdir('tree/unexecutable_dir')
+        text_file('tree/unexecutable_dir/text')
+        os.chmod('tree/unexecutable_dir', 0o600)
+        os.mkdir('tree/totally_unusable_dir')
+        text_file('tree/totally_unusable_dir/text')
+        os.chmod('tree/totally_unusable_dir', 0o100)
 
 
 def ensure_deletability(arg, dirname, fnames):
@@ -176,17 +203,22 @@ def teardown():
                        'text.dont_skip_ext', 'dir.skip_ext', 'skip_dir',
                        'fake_skip_dir', 'text#', 'foo.bar.baz']
     for filename in files_to_delete:
-        try:
-            if os.path.islink(filename) or os.path.isfile(filename):
-                os.unlink(filename)
-            else:
-                os.rmdir(filename)
-        except OSError as e:
-            print('Could not delete %s: %s' % (filename, e), file=sys.stderr)
-    os.unlink('socket_test')
+        if os.path.exists(filename):
+            try:
+                if os.path.islink(filename) or os.path.isfile(filename):
+                    os.unlink(filename)
+                else:
+                    shutil.rmtree(filename)
+            except OSError as e:
+                print('Could not delete %s: %s' % (filename, e), file=sys.stderr)
+    if not sys.platform == 'win32':
+        os.unlink('socket_test')
     os.walk('tree', ensure_deletability, None)
     shutil.rmtree('tree')
 
+    os.chdir(_saved_working_dir)
+    shutil.rmtree(_temp_test_dir)
+    
 
 def test_binary():
     fr = FileRecognizer()
@@ -228,6 +260,7 @@ def test_binary_middle():
     nt.assert_equal(fr.recognize('binary_middle'), 'binary')
 
 
+@skipIf(sys.platform == 'win32', 'no AF_UNIX on win32')
 def test_socket():
     fr = FileRecognizer()
     nt.assert_equal(fr.recognize('socket_test'), 'skip')
@@ -239,6 +272,8 @@ def test_dir():
     nt.assert_equal(fr.recognize('dir'), 'directory')
 
 
+@skipIf(sys.version_info < (3, 3) and sys.platform == 'win32',
+        'Python 3.2+ is required for symlinks on Windows')
 def test_skip_symlinks():
     fr = FileRecognizer(skip_symlink_files=True, skip_symlink_dirs=True)
     nt.assert_equal(fr.recognize('binary_link'), 'link')
@@ -249,6 +284,8 @@ def test_skip_symlinks():
     nt.assert_equal(fr.recognize_directory('dir_link'), 'link')
 
 
+@skipIf(sys.version_info < (3, 3) and sys.platform == 'win32',
+        'Python 3.2+ is required for symlinks on Windows')
 def test_do_not_skip_symlinks():
     fr = FileRecognizer(skip_symlink_files=False, skip_symlink_dirs=False)
     nt.assert_equal(fr.recognize('binary_link'), 'binary')
@@ -267,12 +304,15 @@ def test_skip_hidden():
     nt.assert_equal(fr.recognize_file('.text'), 'skip')
     nt.assert_equal(fr.recognize('.dir'), 'skip')
     nt.assert_equal(fr.recognize_directory('.dir'), 'skip')
-    nt.assert_equal(fr.recognize('.binary_link'), 'skip')
-    nt.assert_equal(fr.recognize_file('.binary_link'), 'skip')
-    nt.assert_equal(fr.recognize('.text_link'), 'skip')
-    nt.assert_equal(fr.recognize_file('.text_link'), 'skip')
-    nt.assert_equal(fr.recognize('.dir_link'), 'skip')
-    nt.assert_equal(fr.recognize_directory('.dir_link'), 'skip')
+
+    if not sys.platform == 'win32' or sys.version_info > (3, 2):
+        nt.assert_equal(fr.recognize('.binary_link'), 'skip')
+        nt.assert_equal(fr.recognize_file('.binary_link'), 'skip')
+        nt.assert_equal(fr.recognize('.text_link'), 'skip')
+        nt.assert_equal(fr.recognize_file('.text_link'), 'skip')
+        nt.assert_equal(fr.recognize('.dir_link'), 'skip')
+        nt.assert_equal(fr.recognize_directory('.dir_link'), 'skip')
+
     nt.assert_equal(fr.recognize('.text.gz'), 'skip')
     nt.assert_equal(fr.recognize_file('.text.gz'), 'skip')
     nt.assert_equal(fr.recognize('.binary.gz'), 'skip')
@@ -298,6 +338,8 @@ def test_skip_weird_exts():
     nt.assert_equal(fr.recognize_file('foo.bar.baz'), 'skip')
 
 
+@skipIf(sys.version_info < (3, 3) and sys.platform == 'win32',
+        'Python 3.2+ is required for symlinks on Windows')
 def test_do_not_skip_hidden_or_symlinks():
     fr = FileRecognizer(skip_hidden_files=False, skip_hidden_dirs=False,
                         skip_symlink_dirs=False, skip_symlink_files=False)
@@ -319,6 +361,8 @@ def test_do_not_skip_hidden_or_symlinks():
     nt.assert_equal(fr.recognize_file('.binary.gz'), 'binary')
 
 
+@skipIf(sys.version_info < (3, 3) and sys.platform == 'win32',
+        'Python 3.2+ is required for symlinks on Windows')
 def test_do_not_skip_hidden_but_skip_symlinks():
     fr = FileRecognizer(skip_hidden_files=False, skip_hidden_dirs=False,
                         skip_symlink_dirs=True, skip_symlink_files=True)
@@ -340,6 +384,7 @@ def test_do_not_skip_hidden_but_skip_symlinks():
     nt.assert_equal(fr.recognize_file('.binary.gz'), 'binary')
 
 
+@skipIf(sys.platform == 'win32', 'Unix permissions model is not supported')
 def test_lack_of_permissions():
     fr = FileRecognizer()
     nt.assert_equal(fr.recognize('unreadable_file'), 'unreadable')
@@ -353,6 +398,7 @@ def test_lack_of_permissions():
                     'directory')
 
 
+@skipIf(sys.platform == 'win32', 'Unix permissions model is not supported')
 def test_symlink_src_unreadable():
     fr = FileRecognizer(skip_symlink_files=False, skip_symlink_dirs=False)
     nt.assert_equal(fr.recognize('unreadable_file_link'), 'unreadable')
@@ -399,7 +445,10 @@ def test_walking():
         ('tree/text', 'text'),
         ('tree/text.dont_skip_ext', 'text'),
     ]
-    result = sorted(fr.walk('tree'))
+    result = []
+    for p, t in sorted(fr.walk('tree')):
+        result.append((p.replace(os.sep, '/'), t))
+
     nt.assert_equal(result, truth)
 
 
@@ -444,12 +493,12 @@ def test_dot_dot():
                         skip_exts=set(['.skip_ext']),
                         skip_dirs=set(['skip_dir']))
     truth = [
-        (os.path.join(os.curdir, 'binary'), 'binary'),
-        (os.path.join(os.curdir, 'dir.skip_ext', 'text'), 'text'),
-        (os.path.join(os.curdir, 'dir', 'subdir', 'text'), 'text'),
-        (os.path.join(os.curdir, 'dir', 'text'), 'text'),
-        (os.path.join(os.curdir, 'text'), 'text'),
-        (os.path.join(os.curdir, 'text.dont_skip_ext'), 'text'),
+        (os.path.join(os.pardir, 'binary'), 'binary'),
+        (os.path.join(os.pardir, 'dir.skip_ext', 'text'), 'text'),
+        (os.path.join(os.pardir, 'dir', 'subdir', 'text'), 'text'),
+        (os.path.join(os.pardir, 'dir', 'text'), 'text'),
+        (os.path.join(os.pardir, 'text'), 'text'),
+        (os.path.join(os.pardir, 'text.dont_skip_ext'), 'text'),
     ]
     result = sorted(fr.walk(os.pardir))
     nt.assert_equal(result, truth)
